searchResults moved to useEffect
