+++ content/src/scripts/containers/SearchBar/SearchBar.jsx
@@ -3,11 +3,11 @@ import { connect } from "react-redux";
 import * as fzy from "fzy.js";
 
 import Backdrop from "../../components/Backdrop";
+import SearchBarResult from "./SearchBarResult";
 import { getSearchTerms } from "../../../../../event/src/actions/API";
 import { fetchURLDetails } from "../../utils/url";
 
 import "./styles.css";
-import fileIcons from "../../utils/file-icons";
 
 function getSearchResults(searchTerms, URLDetails, query) {
   if (
@@ -40,82 +40,6 @@ function getSearchResults(searchTerms, URLDetails, query) {
   return [];
 }
 
-function SearchBarResult({
-  index,
-  term,
-  query,
-  activeResult,
-  setActiveResult,
-}) {
-  let fileLocation = term.split("/");
-  let fileName = fileLocation.splice(-1);
-  let resultClass =
-    index === activeResult
-      ? "spantree-search-result spantree-result-active"
-      : "spantree-search-result";
-  query = query.replace(/ /g, "");
-  const getAlternatingArray = (arr) => {
-    let l = 1;
-    let i = 1;
-    let res = [arr[0]];
-    while (i <= arr.length - 1) {
-      let diff = arr[i] - arr[i - 1];
-      if (diff === 1) {
-        l++;
-      } else {
-        res.push(l);
-        res.push(diff - 1);
-        l = 1;
-      }
-      i++;
-    }
-    res.push(l);
-    return res;
-  };
-  let charLocations = [];
-  let alternatingArray = [];
-  if (query.length !== 0) {
-    charLocations = fzy.positions(query, term);
-    alternatingArray = getAlternatingArray(charLocations);
-  }
-  let isFzy = false;
-  return (
-    <div
-      className={resultClass}
-      onClick={() => {
-        console.log(fileName);
-      }}
-      onMouseEnter={() => {
-        setActiveResult(index);
-      }}
-    >
-      <div className="spantree-search-file">
-        <div className="file-icon">
-          <i className={fileIcons.getClassWithColor(fileName, false)}></i>
-        </div>
-        <div className="spantree-search-filename">{fileName}</div>
-      </div>
-      <div className="spantree-search-filelocation">
-        {alternatingArray.map((len, index) => {
-          let charClass = "";
-          if (isFzy) {
-            charClass = "in-fzy";
-          }
-          isFzy = !isFzy;
-          const currString = term.substr(0, len);
-          term = term.substr(len);
-          return (
-            <span key={index} className={charClass}>
-              {currString}
-            </span>
-          );
-        })}
-        <span className="search-term">{term}</span>
-      </div>
-    </div>
-  );
-}
-
 function SearchBar({
   reloading,
   setReloading,
@@ -125,12 +49,14 @@ function SearchBar({
 }) {
   const [showSearchbar, _setShowSearchbar] = useState(false);
   const showSearchbarRef = useRef(showSearchbar);
+  const [searchResults, setSearchResults] = useState([]);
+  const [searchFor, setSearchFor] = useState("");
+  const [activeResult, setActiveResult] = useState(0);
+
   const setShowSearchbar = (data) => {
     showSearchbarRef.current = data;
     _setShowSearchbar(data);
   };
-  const [searchFor, setSearchFor] = useState("");
-  const [activeResult, setActiveResult] = useState(0);
 
   useEffect(() => {
     const URLDetails = fetchURLDetails();
@@ -142,6 +68,12 @@ function SearchBar({
     });
     document.addEventListener("keydown", handleKeyDown);
   }, []);
+
+  useEffect(() => {
+    setSearchResults(
+      getSearchResults(searchTerms, fetchURLDetails(), searchFor)
+    );
+  }, [searchFor]);
 
   const isMac = ["Macintosh", "MacIntel", "MacPPC", "Mac68K"].reduce(
     (accumulator, currentValue) => {
@@ -193,23 +125,21 @@ function SearchBar({
           />
         </div>
         <div className="spantree-search-results">
-          {getSearchResults(searchTerms, fetchURLDetails(), searchFor).map(
-            (resultTerm, index) => {
-              return (
-                <SearchBarResult
-                  key={index}
-                  index={index}
-                  query={searchFor}
-                  term={resultTerm}
-                  activeResult={activeResult}
-                  setActiveResult={setActiveResult}
-                />
-              );
-            }
-          )}
-          <div className="only-top-result">
+          {searchResults.map((resultTerm, index) => {
+            return (
+              <SearchBarResult
+                key={index}
+                index={index}
+                query={searchFor}
+                term={resultTerm}
+                activeResult={activeResult}
+                setActiveResult={setActiveResult}
+              />
+            );
+          })}
+          {/* <div className="only-top-result">
             Showing only the top 100 results
-          </div>
+          </div> */}
         </div>
       </div>
     </Fragment>
